Getting changes ready for TanukiKit pull request
